requirements.txt
numpy~=1.24.1
h5py~=3.8.0
matplotlib~=3.7.1
opencv-python
torch~=1.13.1+cu116
torchvision~=0.14.1+cu116

pillow~=9.3.0
pandas~=2.0.0
argparse~=1.4.0
tensorboardx~=2.6
thop~=0.1.1-2209072238
tqdm